+++ ui/src/pages/Colors/ColorPicker/constant.ts
@@ -2,7 +2,18 @@ import { FormatType } from "./type";
 
 const formatLabels = ["HEX", "RGB", "RGBA", "HSL", "HSLA"] as const;
 
+const extendedFormatLabels = [
+	...formatLabels.slice(0, 1),
+	"HEX8",
+	...formatLabels.slice(1),
+] as const;
+
 const DATA_OPTIONS = formatLabels.map((label) => ({
+	value: label.toLowerCase() as FormatType,
+	label,
+}));
+
+const EXTENDED_DATA_OPTIONS = extendedFormatLabels.map((label) => ({
 	value: label.toLowerCase() as FormatType,
 	label,
 }));
@@ -10,4 +21,11 @@ const DATA_OPTIONS = formatLabels.map((label) => ({
 const INITIAL_COLOR = "#ffffff";
 const INITIAL_FORMAT = DATA_OPTIONS[0].value;
 
-export { formatLabels, INITIAL_COLOR, INITIAL_FORMAT, DATA_OPTIONS };
+export {
+	formatLabels,
+	extendedFormatLabels,
+	INITIAL_COLOR,
+	INITIAL_FORMAT,
+	DATA_OPTIONS,
+	EXTENDED_DATA_OPTIONS,
+};